refactor: added return value

+++ items/weapon/fists.py
@@ -7,4 +7,4 @@ class Fists:
     MIN_DAMAGE: int = 1
     MAX_DAMAGE: int = 2
     CRITICAL_STRIKE_CHANCE: float = 0.05
-    ACCURACY: int = 0.8
+    ACCURACY: float = 0.8
